REfactored styles into seperate pages and moved redundant code.
