+++ components/displayMessages/displayMessages.styles.tsx
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material/styles'
+
+const StyledFinalMessagesContainer = styled('div')({
+  marginTop: 50,
+  marginBottom: 50
+})
+
+export { StyledFinalMessagesContainer }
